Added manual mode
Added manual mode if want to start yourself pomodoro
