+++ pomodoro/index.html
@@ -76,6 +76,9 @@
         <button class="pomodoro-btn pomodoro-btn-sound" id="sound-toggle">
           <i class="fas fa-volume-up"></i>
         </button>
+        <button class="pomodoro-btn pomodoro-btn-auto" id="auto-start-toggle">
+          <i class="fas fa-forward"></i> Auto
+        </button>
       </div>
     </div>
