+++ pomodoro/js/tasks.js
@@ -13,7 +13,7 @@ const taskContainer = document.querySelector(".task-container > ul");
  * @returns {string} - Escaped text safe for HTML insertion
  */
 function escapeHtml(text) {
-  const div = document.createElement('div');
+  const div = document.createElement("div");
   div.textContent = text;
   return div.innerHTML;
 }
@@ -25,39 +25,41 @@ function escapeHtml(text) {
 if (taskForm) {
   taskForm.addEventListener("submit", (e) => {
     e.preventDefault();
-    
+
     try {
       const taskText = taskInput.value.trim();
       const pomodoroCountRaw = pomodoroCountInput.value.trim();
       const pomodoroCount = parseFloat(pomodoroCountRaw);
-      
+
       if (!taskText) {
         alert("Please enter a task name.");
         return;
       }
-      
+
       if (isNaN(pomodoroCount) || pomodoroCount < 0.1) {
         alert("Please enter a valid number of pomodoros (minimum 0.1).");
         return;
       }
-      
+
       if (Math.round(pomodoroCount * 10) / 10 !== pomodoroCount) {
-        alert("Pomodoros must be in increments of 0.1 (e.g., 0.1, 0.5, 1, 1.5).");
+        alert(
+          "Pomodoros must be in increments of 0.1 (e.g., 0.1, 0.5, 1, 1.5).",
+        );
         return;
       }
-      
+
       const li = document.createElement("li");
       li.className = "tasks";
       li.setAttribute("data-pomodoros", pomodoroCount);
       li.innerHTML = `<span class="task-info"><span class="task-text">${escapeHtml(taskText)}</span> | <span class="task-count">${pomodoroCount}</span> pomodoro${pomodoroCount > 1 ? "s" : ""}</span><i class="fas fa-trash delete-icon"></i>`;
-      
+
       if (taskContainer) {
         taskContainer.appendChild(li);
       }
-      
+
       taskInput.value = "";
       pomodoroCountInput.value = "";
-      
+
       saveTasks();
     } catch (error) {
       console.error("Failed to add task:", error);
@@ -73,16 +75,16 @@ if (taskForm) {
 function decrementFirstTask() {
   try {
     if (!taskContainer) return;
-    
+
     const firstTask = taskContainer.querySelector(".tasks");
     if (!firstTask) return;
-    
+
     let count = parseFloat(firstTask.getAttribute("data-pomodoros"));
     if (isNaN(count)) return;
-    
+
     count -= 0.1;
     count = Math.round(count * 10) / 10;
-    
+
     if (count <= 0) {
       firstTask.remove();
     } else {
@@ -94,7 +96,7 @@ function decrementFirstTask() {
         countSpan.nextSibling.textContent = textAfterCount;
       }
     }
-    
+
     saveTasks();
   } catch (error) {
     console.error("Failed to decrement task:", error);
