+++ resume/about.html
@@ -25,7 +25,7 @@
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.0/font/bootstrap-icons.css"
     />
-    <link href="styles.css" rel="stylesheet" />
+    <link href="css/styles.css" rel="stylesheet" />
   </head>
   <body>
     <!-- Navbar Container -->
@@ -66,15 +66,14 @@
                   </div>
                   <p class="card-text">
                     I began my career in the culinary arts, obtaining my certification from 
-                    Fitjar Vidaregåande Skule. This culinary foundation laid the groundwork for 
-                    my detail-oriented approach and ability to perform effectively in high-pressure 
-                    environments.
+                    Fitjar Vidaregåande Skule. This culinary foundation provided the groundwork for 
+                    my detail-oriented approach and ability to perform effectively under pressure.
                   </p>
                   <p class="card-text">
                     My career evolved as I moved into industrial roles, working at Nova Sea, 
                     Bremnes Seashore, and as a cleaner and fire watch at Subshore AS on the 
-                    Njord project. These positions have honed my skills in safety protocols, 
-                    precision work, and collaboration.
+                    Njord A project. These positions have honed my skills in safety protocols, 
+                    precision work, and team collaboration.
                   </p>
                   <p class="card-text">
                     <i class="bi bi-envelope-fill me-2"></i>
@@ -267,7 +266,7 @@
                       <ul class="list-unstyled">
                         <li class="mb-3">
                           <i class="bi bi-check-circle-fill text-success me-2"></i>
-                          Further develop industrial expertise and certifications
+                          Expand industrial expertise and obtain additional certifications
                         </li>
                         <li class="mb-3">
                           <i class="bi bi-check-circle-fill text-success me-2"></i>
@@ -390,6 +389,11 @@
 
     <!-- Scripts -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="common.js"></script>
+    <script src="js/utils.js"></script>
+    <script src="js/navbar.js"></script>
+    <script src="js/footer.js"></script>
+    <script src="js/profile.js"></script>
+    <script src="js/darkmode.js"></script>
+    <script src="js/common.js"></script>
   </body>
 </html>
