+++ resume/css/cards.css
@@ -1,0 +1,185 @@
+/* Cards */
+.card {
+  background-color: var(--card-bg);
+  border: none;
+  border-radius: 1rem;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  transition: transform var(--transition-speed), box-shadow var(--transition-speed);
+}
+
+.card:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 12px rgba(0, 0, 0, 0.15);
+}
+
+.card-body {
+  padding: 1.5rem;
+}
+
+/* Project Cards */
+.project-card {
+  text-align: center;
+  padding: 2rem 1rem;
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+  border: none;
+  border-radius: 15px;
+  overflow: hidden;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  height: 100%;
+}
+
+.project-card:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 15px rgba(0, 0, 0, 0.2);
+}
+
+.project-card .card-body {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.project-card:hover .project-image-container img {
+  transform: scale(1.05);
+}
+
+/* Profile Card */
+.profile-card {
+  border: none;
+  border-radius: 1rem;
+  background-color: var(--card-bg);
+  transition: all var(--transition-speed);
+}
+
+.profile-card img {
+  border: 4px solid var(--primary-color);
+  padding: 4px;
+  transition: all var(--transition-speed);
+}
+
+.profile-card:hover img {
+  transform: scale(1.05);
+}
+
+/* Timeline Cards */
+.timeline-card {
+  height: 100%;
+  transition: all var(--transition-speed);
+  border: none;
+  border-radius: 1rem;
+  background-color: var(--card-bg);
+  position: relative;
+  overflow: hidden;
+}
+
+.timeline-card::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  height: 4px;
+  background: linear-gradient(90deg, var(--primary-color), #0099ff);
+}
+
+.timeline-card:hover {
+  transform: translateY(-10px);
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
+
+.timeline-card .card-body {
+  padding: 2rem 1.5rem;
+}
+
+.timeline-card h4 {
+  color: var(--primary-color);
+  margin-bottom: 1rem;
+}
+
+.timeline-card ul li {
+  margin-bottom: 0.5rem;
+  color: var(--text-color);
+}
+
+.timeline-card i {
+  transition: transform var(--transition-speed);
+}
+
+.timeline-card:hover i {
+  transform: scale(1.1);
+}
+
+/* Interactive Cards */
+.interactive-card {
+  border: none;
+  border-radius: 1rem;
+  background-color: var(--card-bg);
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  transition: all var(--transition-speed);
+  overflow: hidden;
+}
+
+.interactive-card:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
+
+/* Social Cards */
+.social-card {
+  height: 100%;
+  border: none;
+  border-radius: 1rem;
+  background-color: var(--card-bg);
+  transition: all var(--transition-speed);
+  overflow: hidden;
+}
+
+.social-card::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  height: 4px;
+  background: linear-gradient(90deg, var(--primary-color), #0099ff);
+}
+
+.social-card:hover {
+  transform: translateY(-10px);
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
+
+.social-card i {
+  transition: transform var(--transition-speed);
+}
+
+.social-card:hover i {
+  transform: scale(1.1);
+}
+
+/* Contact Card */
+.contact-card {
+  border: none;
+  border-radius: 1rem;
+  background-color: var(--card-bg);
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  transition: all var(--transition-speed);
+}
+
+.contact-card:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
+
+/* About Card */
+.about-card {
+  border: none;
+  border-radius: 1rem;
+  background-color: var(--card-bg);
+  transition: all var(--transition-speed);
+}
+
+.about-card:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
