+++ resume/css/components.css
@@ -1,0 +1,224 @@
+/* Skills Section */
+.skill-item {
+  padding: 2rem;
+  border-radius: 1rem;
+  background-color: var(--card-bg);
+  transition: transform var(--transition-speed);
+  height: 100%;
+}
+
+.skill-item:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
+
+.skill-item i {
+  transition: transform var(--transition-speed);
+}
+
+.skill-item:hover i {
+  transform: scale(1.1);
+}
+
+/* Footer */
+.footer {
+  background-color: var(--navbar-bg);
+  color: #fff;
+}
+
+/* Education Section */
+.education-item {
+  position: relative;
+  padding-left: 2rem;
+}
+
+.education-item::before {
+  content: '';
+  position: absolute;
+  left: 0;
+  top: 0;
+  bottom: 0;
+  width: 2px;
+  background-color: var(--primary-color);
+}
+
+/* About Page Sections */
+.about-section {
+  position: relative;
+  overflow: hidden;
+}
+
+.about-section::after {
+  content: '';
+  position: absolute;
+  bottom: -50%;
+  right: -50%;
+  width: 100%;
+  height: 100%;
+  background: radial-gradient(circle, rgba(13, 110, 253, 0.1) 0%, rgba(13, 110, 253, 0) 70%);
+  z-index: -1;
+}
+
+/* Goals Section */
+.goals-list li {
+  padding: 1rem;
+  margin-bottom: 1rem;
+  border-radius: 0.5rem;
+  background-color: var(--card-bg);
+  transition: all var(--transition-speed);
+}
+
+.goals-list li:hover {
+  transform: translateX(10px);
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+/* Contact Info */
+.contact-info {
+  padding: 2rem;
+  border-radius: 1rem;
+  background-color: var(--card-bg);
+  transition: all var(--transition-speed);
+}
+
+.contact-info:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+}
+
+.contact-info i {
+  transition: transform var(--transition-speed);
+}
+
+.contact-info:hover i {
+  transform: scale(1.1);
+}
+
+/* Project Image Styles */
+.project-image-container {
+  position: relative;
+  overflow: hidden;
+  aspect-ratio: 16/9;
+}
+
+.project-image-container img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  transition: transform 0.3s ease;
+}
+
+.image-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  opacity: 0;
+  transition: opacity 0.3s ease;
+}
+
+.project-card:hover .image-overlay {
+  opacity: 1;
+}
+
+/* Badges */
+.badge {
+  font-size: 0.8rem;
+  padding: 0.5em 0.8em;
+  margin-bottom: 0.5rem;
+  border-radius: 20px;
+}
+
+/* Project Filters */
+#projectFilters {
+  margin-bottom: 2rem;
+}
+
+#projectFilters .btn {
+  border-radius: 20px;
+  padding: 0.5rem 1.5rem;
+  font-weight: 500;
+  transition: all 0.3s ease;
+}
+
+#projectFilters .btn:hover,
+#projectFilters .btn.active {
+  background-color: var(--primary-color);
+  color: white;
+  transform: translateY(-2px);
+}
+
+/* Carousel */
+.carousel-item img {
+  border-radius: 10px;
+  object-fit: contain;
+  max-height: 70vh;
+}
+
+.carousel-indicators-container {
+  display: flex;
+  justify-content: center;
+  gap: 0.5rem;
+}
+
+.carousel-indicators-container .btn {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-weight: 500;
+}
+
+/* Modal */
+.modal-content {
+  border-radius: 15px;
+  border: none;
+}
+
+.modal-header {
+  border-bottom: none;
+  padding: 1.5rem;
+}
+
+.modal-body {
+  padding: 1.5rem;
+}
+
+/* Project Description */
+.project-description {
+  position: relative;
+  margin-bottom: 1rem;
+}
+
+.description-text {
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  margin-bottom: 0.5rem;
+  transition: all 0.3s ease;
+}
+
+.description-text.expanded {
+  -webkit-line-clamp: unset;
+  max-height: none;
+}
+
+.show-more-btn {
+  color: var(--primary-color);
+  text-decoration: none;
+  font-weight: 500;
+  transition: all 0.3s ease;
+}
+
+.show-more-btn:hover {
+  color: var(--secondary-color);
+  text-decoration: underline;
+}
