+++ resume/css/navigation.css
@@ -1,0 +1,37 @@
+/* Navigation */
+.navbar {
+  background-color: var(--navbar-bg);
+  transition: all var(--transition-speed);
+  padding: 1rem 0;
+}
+
+.navbar-scrolled {
+  padding: 0.5rem 0;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+.navbar-brand {
+  font-weight: 700;
+  font-size: 1.5rem;
+}
+
+.nav-link {
+  font-weight: 500;
+  padding: 0.5rem 1rem !important;
+  transition: color var(--transition-speed);
+}
+
+/* Dark Mode Toggle */
+#darkModeToggle {
+  padding: 0.5rem;
+  border-radius: 50%;
+  width: 2.5rem;
+  height: 2.5rem;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+}
+
+#darkModeToggle i {
+  font-size: 1.2rem;
+}
