+++ resume/css/styles.css
@@ -1,0 +1,12 @@
+/* Main Stylesheet - Imports all component stylesheets */
+@import url('variables.css');
+@import url('base.css');
+@import url('navigation.css');
+@import url('hero.css');
+@import url('animations.css');
+@import url('cards.css');
+@import url('buttons.css');
+@import url('forms.css');
+@import url('tools.css');
+@import url('components.css');
+@import url('responsive.css');
