+++ resume/index.html
@@ -25,7 +25,7 @@
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.0/font/bootstrap-icons.css"
     />
-    <link href="styles.css" rel="stylesheet" />
+    <link href="css/styles.css" rel="stylesheet" />
   </head>
   <body>
     <!-- Navbar Container -->
@@ -87,7 +87,7 @@
                   <i class="bi bi-fire display-4 text-primary mb-3"></i>
                   <h3 class="card-title">Safety Operations</h3>
                   <p class="card-text">
-                    Fire watch and cleaning roles at Subshore AS on the Njord project with relevant certifications.
+                    Fire watch and cleaning roles at Subshore AS on the Njord A project with relevant certifications.
                   </p>
                   <a href="resume.html#certifications" class="btn btn-outline-primary"
                     >See Certifications</a
@@ -223,6 +223,11 @@
 
     <!-- Bootstrap JS and other scripts -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="common.js"></script>
+    <script src="js/utils.js"></script>
+    <script src="js/navbar.js"></script>
+    <script src="js/footer.js"></script>
+    <script src="js/profile.js"></script>
+    <script src="js/darkmode.js"></script>
+    <script src="js/common.js"></script>
   </body>
 </html>
