+++ resume/js/common.js
@@ -1,0 +1,19 @@
+/**
+ * Main Common Module
+ * Initializes all components and modules
+ */
+
+document.addEventListener('DOMContentLoaded', function() {
+    // Inject navbar and footer
+    injectNavbar();
+    injectFooter();
+    
+    // Initialize navbar scroll behavior
+    initializeNavbarScroll();
+    
+    // Randomize profile pictures
+    randomizeProfilePictures();
+    
+    // Initialize dark mode
+    initializeDarkMode();
+});
