+++ resume/resume.html
@@ -39,7 +39,7 @@
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.0/font/bootstrap-icons.css"
     />
-    <link href="styles.css" rel="stylesheet" />
+    <link href="css/styles.css" rel="stylesheet" />
   </head>
   <body>
     <!-- Navbar Container -->
@@ -262,7 +262,7 @@
                         </div>
                         <div class="timeline-content">
                           <h5>Nova Sea</h5>
-                          <p>Full-time position performing various roles including sløyehald, washing, cooling, and packaging tasks</p>
+                          <p>Full-time position performing various roles including gutting station, washing, cooling, and packaging tasks</p>
                         </div>
                       </div>
                     </div>
@@ -487,6 +487,11 @@
 
     <!-- Scripts -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="common.js"></script>
+    <script src="js/utils.js"></script>
+    <script src="js/navbar.js"></script>
+    <script src="js/footer.js"></script>
+    <script src="js/profile.js"></script>
+    <script src="js/darkmode.js"></script>
+    <script src="js/common.js"></script>
   </body>
 </html>
